Code done, now time for practical part
"funciones.py" deleted because it didn't gave any real value just having
that function there by itself.

The thing that now has to be done is the theoretical part and start
changing the values for the respective elements used in those
experiments.

Also, this repo needs a "README.md" that has all of this info.