Putting all together, theory and practical parts
After adding the videos and taking all the data of the angles for the
pendulums by hand, there was the need to make the graphs to be able to
compare them in general.

The coefficient c had to be found by try and error because the various
methods that were proven to found it were useless and didn't work.